Better separation of responsibilities for share button feature.
