+++ CHANGELOG.md
@@ -26,7 +26,8 @@ Firebase and static code analyzers too.
 - New ':commons_file' and ':commons_android' modules.
 - Changelog formatting.
 - Default language to English.
-- Gradle version to [v5.6.2](https://docs.gradle.org/5.6.2/release-notes.html)
+- Gradle version to [v5.6.2](https://docs.gradle.org/5.6.2/release-notes.html).
+- Better separation of responsibilities.
 
 ### Fixed
 - Many SCA suggestions applied.
